+++ src/command_modules/azure-cli-sqlvm/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+0.1.2
++++++
+* Now customer can set sql image sku type when creating or updating a sql vm.
+* If sql server license type has not been provided, create command will infer it from Compute's image offering.
+
 0.1.1
 +++++
 * Minor fixes.